+++ packages/alfa-dom/src/node/predicate/has-inclusive-descendant.ts
@@ -1,14 +1,14 @@
 import type { Predicate } from "@siteimprove/alfa-predicate";
 
 import type { Node } from "../index.ts";
+import { BaseNode } from "../node.js";
 
 /**
  * @public
  */
 export function hasInclusiveDescendant(
-  defaultTraversal: Node.Traversal,
-): (predicate: Predicate<Node>, options?: Node.Traversal) => Predicate<Node> {
-  return (predicate, options = defaultTraversal) =>
-    (node) =>
-      node.inclusiveDescendants(options).some(predicate);
+  predicate: Predicate<Node>,
+  options: Node.Traversal = BaseNode.Traversal.empty,
+): Predicate<Node> {
+  return (node) => node.inclusiveDescendants(options).some(predicate);
 }