+++ packages/alfa-dom/src/node/predicate/has-text-content.ts
@@ -2,6 +2,7 @@ import { Iterable } from "@siteimprove/alfa-iterable";
 import { Predicate } from "@siteimprove/alfa-predicate";
 
 import type { Node } from "../index.ts";
+import { BaseNode } from "../node.js";
 
 const { isEmpty } = Iterable;
 const { not } = Predicate;
@@ -10,12 +11,8 @@ const { not } = Predicate;
  * @public
  */
 export function hasTextContent(
-  defaultTraversal: Node.Traversal,
-): (
-  predicate?: Predicate<string>,
-  options?: Node.Traversal,
-) => Predicate<Node> {
-  return (predicate = not(isEmpty), options = defaultTraversal) =>
-    (node) =>
-      predicate(node.textContent(options));
+  predicate: Predicate<string> = not(isEmpty),
+  options: Node.Traversal = BaseNode.Traversal.empty,
+): Predicate<Node> {
+  return (node) => predicate(node.textContent(options));
 }