+++ packages/alfa-dom/src/node/query/element-id-map.ts
@@ -1,7 +1,7 @@
 import { Cache } from "@siteimprove/alfa-cache";
 import { Map } from "@siteimprove/alfa-map";
 
-import type { BaseNode } from "../node.ts";
+import type { Node } from "../index.ts";
 import { Document } from "../document.ts";
 import type { Element } from "../slotable/element.ts";
 
@@ -18,7 +18,7 @@ const elementMapCache = Cache.empty<Document, Map<string, Element>>();
  *
  * @public
  */
-export function getElementIdMap(node: BaseNode): Map<string, Element> {
+export function getElementIdMap(node: Node): Map<string, Element> {
   if (Document.isDocument(node)) {
     return elementMapCache.get(node, () => buildElementIdMap(node));
   }
@@ -26,7 +26,7 @@ export function getElementIdMap(node: BaseNode): Map<string, Element> {
   return buildElementIdMap(node);
 }
 
-function buildElementIdMap(node: BaseNode): Map<string, Element> {
+function buildElementIdMap(node: Node): Map<string, Element> {
   // Build a map from ID -> element to allow fast resolution of ID references.
   // The collected references are added to the map in reverse order to ensure
   // that the first occurrence of a given ID is what ends up in the map in