+++ packages/alfa-dom/src/node/query/index.ts
@@ -1,7 +1,7 @@
 import * as descendants from "./descendants.ts";
 import * as elementIdMap from "./element-id-map.ts";
 
-import type { BaseNode } from "../node.ts";
+import type { Node } from "../index.ts";
 
 /**
  * @public
@@ -15,6 +15,6 @@ export namespace Query {
   export const getElementIdMap = elementIdMap.getElementIdMap;
 
   export type TextGroup = descendants.TextGroup;
-  export type TextGroupOptions<N extends BaseNode = BaseNode> =
+  export type TextGroupOptions<N extends Node = Node> =
     descendants.TextGroupOptions<N>;
 }