+++ packages/alfa-dom/test/node/traversal/get-nodes-between.spec.tsx
@@ -1,9 +1,7 @@
 import { test } from "@siteimprove/alfa-test";
 
-import { h, Node } from "../../../dist/index.js";
-import { getNodesBetween as GNB } from "../../../dist/node/traversal/get-nodes-between.js";
-
-const getNodesBetween = GNB(Node.Traversal.empty);
+import { h } from "../../../dist/index.js";
+import { getNodesBetween } from "../../../dist/node/traversal/get-nodes-between.js";
 
 const hello = h.text("Hello");
 const spanHello = h.element("span", [], [hello]);