+++ packages/alfa-collection/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @siteimprove/alfa-collection
 
+## 0.114.0
+
 ## 0.113.0
 
 ## 0.112.0