+++ packages/alfa-css-feature/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @siteimprove/alfa-css-feature
 
+## 0.114.0
+
 ## 0.113.0
 
 ## 0.112.0