+++ packages/alfa-style/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @siteimprove/alfa-style
 
+## 0.114.0
+
 ## 0.113.0
 
 ### Patch Changes